client(GUI)/requirements.txt
PyAudio==0.2.14
python-dotenv==1.0.1